backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.3
django-cors-headers>=4.4
psycopg[binary]>=3.2
celery[redis]>=5.4
gunicorn>=23.0
Pillow>=10.4
redis>=5.2
